requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
asyncio==3.4.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
prisma==0.11.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
